+++ ADUtilsTests/ADViewLayoutTestsObjc.m
@@ -35,26 +35,31 @@
         it(@"Pin to superview", ^{
             [subview ad_pinToSuperview];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(4));
         });
 
         it(@"Pin to superview edges", ^{
             [subview ad_pinToSuperviewWithEdges:UIRectEdgeAll];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(4));
         });
 
         it(@"Pin to superview insets", ^{
             [subview ad_pinToSuperviewWithInsets:UIEdgeInsetsZero];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(4));
         });
 
         it(@"Pin to superview edges insets", ^{
             [subview ad_pinToSuperviewWithEdges:UIRectEdgeAll insets:UIEdgeInsetsZero];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(4));
         });
 
         it(@"Pin to superview edges insets priority", ^{
             [subview ad_pinToSuperviewWithEdges:UIRectEdgeAll insets:UIEdgeInsetsZero priority:UILayoutPriorityRequired];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(4));
         });
 
         // Center in superview
@@ -62,21 +67,25 @@
         it(@"Center in superview", ^{
             [subview ad_centerInSuperview];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(2));
         });
 
         it(@"Center in superview priority", ^{
             [subview ad_centerInSuperviewWithPriority:UILayoutPriorityRequired];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(2));
         });
 
         it(@"Center in superview axis", ^{
             [subview ad_centerInSuperviewAlongAxis:UILayoutConstraintAxisVertical];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(1));
         });
 
         it(@"Center in superview axis priority", ^{
             [subview ad_centerInSuperviewAlongAxis:UILayoutConstraintAxisVertical priority:UILayoutPriorityRequired];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(1));
         });
 
         // Contrain to size
@@ -84,11 +93,13 @@
         it(@"Constrain to size", ^{
             [subview ad_constrainToSize:CGSizeZero];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(subview.constraints.count).to(equal(2));
         });
 
         it(@"Constrain to size priority", ^{
             [subview ad_constrainToSize:CGSizeZero priority:UILayoutPriorityRequired];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(subview.constraints.count).to(equal(2));
         });
 
         // Constrain in superview
@@ -96,26 +107,31 @@
         it(@"Constrain in superview", ^{
             [subview ad_constrainInSuperview];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(4));
         });
 
         it(@"Constrain in superview edges", ^{
             [subview ad_constrainInSuperviewWithEdges:UIRectEdgeAll];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(4));
         });
 
         it(@"Constrain in superview insets", ^{
             [subview ad_constrainInSuperviewWithInsets:UIEdgeInsetsZero];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(4));
         });
 
         it(@"Constrain in superview edges insets", ^{
             [subview ad_constrainInSuperviewWithEdges:UIRectEdgeAll insets:UIEdgeInsetsZero];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(4));
         });
 
         it(@"Constrain in superview edges insets priority", ^{
             [subview ad_constrainInSuperviewWithEdges:UIRectEdgeAll insets:UIEdgeInsetsZero priority:UILayoutPriorityRequired];
             expect(subview.translatesAutoresizingMaskIntoConstraints).to(beFalse());
+            expect(view.constraints.count).to(equal(4));
         });
      });
 }